refactor(api/types): move all params to `/params`
