--- app/api/types/get-top-items-params.ts
@@ -1,8 +1,0 @@
-import { TimeRange } from './time-range'
-import { UsersParams } from './users-params'
-
-export interface GetTopItemsParams extends UsersParams {
-  limit?: number
-  offset?: number
-  timeRange?: TimeRange
-}
