--- app/api/types/users-params.ts
@@ -1,3 +1,0 @@
-export interface UsersParams {
-  userId: string
-}
